# ui/packages/tidb-dashboard-client/swagger/spec.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dashboard API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/dashboard/api",
    "paths": {
...
        "/info/databases": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "List all databases",
                "operationId": "infoListDatabases",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/info/info": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get information about this TiDB Dashboard",
                "operationId": "infoGet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/info.InfoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/info/tables": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "List tables by database name",
                "operationId": "infoListTables",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Database name",
                        "name": "database_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/info.tableSchema"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/info/whoami": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get information about current session",
                "operationId": "infoWhoami",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/info.WhoAmIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "info.InfoResponse": {
            "type": "object",
            "properties": {
                "enable_experimental": {
                    "type": "boolean"
                },
                "enable_telemetry": {
                    "type": "boolean"
                },
                "ngm_state": {
                    "type": "string"
                },
                "no_tidb": {
                    "type": "boolean"
                },
                "supported_features": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "version": {
                    "$ref": "#/definitions/version.Info"
                }
            }
        },
        "info.WhoAmIResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "is_shareable": {
                    "type": "boolean"
                },
                "is_writeable": {
                    "type": "boolean"
                }
            }
        },
        "info.tableSchema": {
            "type": "object",
            "properties": {
                "table_id": {
                    "type": "string"
                },
                "table_name": {
                    "type": "string"
                }
            }
        },
...
        "rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "boolean"
                },
                "full_text": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "version.Info": {
            "type": "object",
            "properties": {
                "build_git_hash": {
                    "type": "string"
                },
                "build_time": {
                    "type": "string"
                },
                "internal_version": {
                    "type": "string"
                },
                "pd_version": {
                    "type": "string"
                },
                "standalone": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}